Refactor: add fcsapi endpoint folder structure
