+++ packages/sources/fcsapi/src/config.ts
@@ -3,10 +3,12 @@ import { Config } from '@chainlink/types'
 
 export const NAME = 'FCSAPI'
 
+export const DEFAULT_ENDPOINT = 'common'
 export const DEFAULT_BASE_URL = 'https://fcsapi.com/api-v3/'
 
 export const makeConfig = (prefix?: string): Config => {
   const config = Requester.getDefaultConfig(prefix, true)
   config.api.baseURL = config.api.baseURL || DEFAULT_BASE_URL
+  config.defaultEndpoint = DEFAULT_ENDPOINT
   return config
 }
