+++ packages/sources/fcsapi/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export * as common from './common'
